+++ src/pub/proxy_edit/general.js
@@ -27,6 +27,13 @@ const log_level_opt = [
     {key: `info`, value: 'info'},
 ];
 
+const proxy_connection_type_opt = [
+    {key: 'Default (HTTP)', value: ''},
+    {key: 'HTTP', value: 'http'},
+    {key: 'HTTPS', value: 'https'},
+    {key: 'SOCKS5', value: 'socks'},
+];
+
 export default class General extends Pure_component {
     state = {};
     get_curr_plan = setdb.get('head.proxy_edit.get_curr_plan');
@@ -83,11 +90,10 @@ export default class General extends Pure_component {
               <Tab_context.Provider value="general">
                 <Config type="text" id="internal_name"/>
                 <Config type="number" id="port"/>
-                <Config type="number" id="socks" disabled={true}
-                  val_id="port"/>
                 <Config type="pins" id="whitelist_ips" disabled={disabled_wl}/>
+                <Config type="select" data={proxy_connection_type_opt}
+                  id="proxy_connection_type"/>
                 <Config type="yes_no" id="ssl" on_change={this.on_change_ssl}/>
-                <Config type="yes_no" id="secure_proxy"/>
                 <Config type="yes_no" id="insecure"
                     disabled={!this.state.form.ssl}/>
                 <Config type="select" data={route_err_opt} id="route_err"/>